+++ windows/include/flutter_volume_controller/volume_controller.h
@@ -19,15 +19,15 @@ namespace flutter_volume_controller {
 
 		void DisposeNotification();
 
-		bool SetVolume(float volume);
+		bool SetVolume(double volume);
 
 		bool SetMaxVolume();
 
 		bool SetMinVolume();
 
-		bool SetVolumeUp(float step);
+		bool SetVolumeUp(double step);
 
-		bool SetVolumeDown(float step);
+		bool SetVolumeDown(double step);
 
 		bool SetVolumeUpBySystemStep();
 
@@ -37,7 +37,7 @@ namespace flutter_volume_controller {
 
 		bool ToggleMute();
 
-		std::optional<float> GetCurrentVolume();
+		std::optional<double> GetCurrentVolume();
 
 		std::optional<bool> GetMute();
